bert_adversarial/requirements.txt
requests
absl-py
gensim
matplotlib
nltk
numpy
pandas
py-params
params-flow
scikit-learn
scipy
sentencepiece
six
spacy
tensorflow-gpu
bert-for-tf2
textblob
tqdm
